feat(ORCH-046): extract AuthSessionRepository and AuthRecord into dedicated module
